Move PAZ reader to its own module
