+++ sourcespec2/input/augment_traces.py
@@ -20,7 +20,7 @@ import logging
 import contextlib
 from obspy.core.util import AttribDict
 from ..setup import config
-from .station_metadata import PAZ
+from .station_metadata_parsers import PAZ
 from .sac_header import (
     compute_sensitivity_from_SAC,
     get_instrument_from_SAC, get_station_coordinates_from_SAC,
